bit of code rewrite
Before this commit the gitcrypt instance had a status flag indicating if
files were encrypted or not. This opens up for error when the status is
marked as encrypted and i add files, and if i decrypt something will go
bad. Now, each file has an attribute indicating if the file is encrypted
or not.. this way when encrypt/decrypt command is executed, only files
encrypted/decrypted will be encrypted/decrypted.

Also added a test for delete files